Refactor `ethrex` initialization
